// client/src/app/interfaces/feedback-request.ts
export class FeedbackRequest {
    maxReviews: number;
    finishedReviews: number;
    outstandingReviews: number;
    creditsOffered: number;
    daysAllowed: number;
    scriptId: string;
    userId: string;
    minimumReviewerRating: number;
    submitTime: number;
    refundableCredits: number;
    untakenReviews: number;
}